requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
fastapi-cors==0.0.6
httpx==0.25.2
pillow==10.1.0
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4